requirements.txt
Flask==1.1.2
google-cloud-texttospeech==2.2.0
requests==2.24.0
spacy==2.3.2
tokenizers==0.9.3
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
transformers==3.5.1
urllib3==1.25.11
progressbar
libmagic
Werkzeug==1.0.1
gunicorn== 19.9.0
python-magic
json
